+++ R/igcop-package.R
@@ -1,0 +1,5 @@
+## usethis namespace: start
+#' @importFrom Rcpp evalCpp
+#' @useDynLib igcop, .registration = TRUE
+## usethis namespace: end
+NULL