+++ tests/testthat/test-log_densities.R
@@ -17,3 +17,4 @@ test_that("exp(log IGL density) returns the IGL density", {
     )
   }
 })
+