+++ tests/testthat/test-numerical_derivatives-igl.R
@@ -6,6 +6,9 @@ digl_gaussian <- function(u, v, alpha) {
 }
 
 test_that("density matches the numerical density obtained from the cdf", {
+  uv <- expand.grid(u = 1:4 / 5, v = 1:4 / 5)
+  .u <- uv$u
+  .v <- uv$v
   #' Function to calculate numerical derivative
   digl_gaussian_numerical <- function(u, v, alpha, eps = 1.e-5) {
     x <- qnorm(u)
@@ -19,11 +22,14 @@ test_that("density matches the numerical density obtained from the cdf", {
   for (alpha_ in .alpha) {
     pdf1 <- digl_gaussian_numerical(.u, .v, alpha = alpha_)
     pdf2 <- digl_gaussian(.u, .v, alpha = alpha_)
-    expect_equal(pdf1, pdf2, tolerance = 1e-4)
+    expect_equal(pdf1, pdf2, tolerance = 1e-3)
   }
 })
 
 test_that("the 2|1 cdf matches the numerically obtained cdf", {
+  uv <- expand.grid(u = 1:4 / 5, v = 1:4 / 5)
+  .u <- uv$u
+  .v <- uv$v
   #' Calculate numerical derivative
   pcondigl21_numerical <- function(v, u, alpha, eps = 1.e-8) {
     cdf11 <- pigl(u, v, alpha = alpha)
@@ -38,8 +44,11 @@ test_that("the 2|1 cdf matches the numerically obtained cdf", {
 })
 
 test_that("the 1|2 cdf matches the numerically obtained cdf", {
+  uv <- expand.grid(u = 1:4 / 5, v = 1:4 / 5)
+  .u <- uv$u
+  .v <- uv$v
   #' Calculate numerical derivative
-  pcondigl12_log_numerical <- function(u, v, alpha, eps = 1.e-9) {
+  pcondigl12_log_numerical <- function(u, v, alpha, eps = 1.e-5) {
     x <- -log(u)
     y <- -log(v)
     cdf11 <- pigl(u, v, alpha = alpha)
@@ -49,14 +58,18 @@ test_that("the 1|2 cdf matches the numerically obtained cdf", {
   for (alpha_ in .alpha) {
     pcond1 <- pcondigl12_log_numerical(.u, .v, alpha = alpha_)
     pcond2 <- pcondigl12(.u, .v, alpha = alpha_) * (-.v)
-    expect_equal(pcond1, pcond2, tolerance = 1e-5)
+    expect_equal(pcond1, pcond2, tolerance = 1e-4)
   }
 })
 
 
-test_that("density matches the numerical density obtained from the 2|1 conditional distribution", {
+test_that("density matches the numerical density obtained from the 2|1
+          conditional distribution", {
+  uv <- expand.grid(u = 1:4 / 5, v = 1:4 / 5)
+  .u <- uv$u
+  .v <- uv$v
   #' Calculate numerical derivative
-  digl_numerical_from_2g1 <- function(u, v, alpha, eps = 1.e-10) {
+  digl_numerical_from_2g1 <- function(u, v, alpha, eps = 1.e-9) {
     conda <- pcondigl21(v, u, alpha = alpha)
     condb <- pcondigl21(v + eps, u, alpha = alpha)
     (condb - conda) / eps
@@ -64,7 +77,7 @@ test_that("density matches the numerical density obtained from the 2|1 condition
   for (alpha_ in .alpha) {
     pdf1 <- digl_numerical_from_2g1(.u, .v, alpha = alpha_)
     pdf2 <- digl(.u, .v, alpha = alpha_)
-    expect_equal(pdf1, pdf2, tolerance = 1e-5)
+    expect_equal(pdf1, pdf2, tolerance = 1e-4)
   }
 })