+++ NEWS.md
@@ -1,3 +1,13 @@
+# igcop 1.0.0
+
+* The package now has a C++ backend, by converting the four Newton-Raphson algorithms and their dependencies from R to C++. 
+
+Smaller changes:
+
+* Renamed `tau` argument to `p` in IG and IGL quantile functions. 
+* Removed `igl_gen_DD` and `igamma()` internal functions, which were not used for copula quantities.
+
+
 # igcop 0.2.0
 
 * Added a `NEWS.md` file to track changes to the package.